@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -28,16 +28,18 @@ dependencies:
     path: ../
 
   # using this to take screenshot
-  screenshot: ^0.0.2
+  screenshot: ^1.1.0
 
   # using this to open default email composed
-  flutter_email_sender: ^2.0.0
+  flutter_email_sender: ^5.0.2
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
+dependency_overrides:
+  plugin_platform_interface: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
