@@ -5,24 +5,28 @@ author: Shakti S.P Swain<shakti.swain.mi@gmail.com>
 homepage: https://github.com/Mindinventory/image-collage-widget.git
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_staggered_grid_view: ^0.3.0
+  flutter_staggered_grid_view: ^0.4.0
   flutter_bloc: ^7.0.1
-  flutter_file_manager: ^0.2.0
-  permission_handler: ^3.2.2
-  image_picker: ^0.6.1+4
-  equatable: ^1.2.6
+#  flutter_file_manager: ^0.2.0
+  flutter_file_utils: ^0.1.1+5
+  permission_handler: ^8.1.4+1
+  image_picker: ^0.8.2
+  equatable: ^2.0.3
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   pedantic: ^1.9.0
+
+dependency_overrides:
+  plugin_platform_interface: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
